+++ src/main/java/KedenApp/dto/PackageKeden.java
@@ -3,12 +3,14 @@ package KedenApp.dto;
 import lombok.Data;
 import lombok.experimental.Accessors;
 
+import java.math.BigDecimal;
+
 @Data
 @Accessors(chain = true)
 public class PackageKeden {
 
     private String commodityCode;
     private String goodsDescriptionText;
-    private Double unifiedGrossMassMeasure;
-    private Double currencyInAmount;
+    private BigDecimal unifiedGrossMassMeasure;
+    private BigDecimal currencyInAmount;
 }
